+++ src/libs/database/mysql/adapters/MySQL2ConnectionAdapter.ts
@@ -7,7 +7,7 @@ import type {
   QueryResult,
   ConnectionConfig,
   PoolStats,
-} from "@/libs/database/interfaces/IDatabaseConnection";
+} from "@/libs/database/interfaces";
 import { MySQL2TransactionAdapter } from "./MySQL2TransactionAdapter";
 
 /**
