+++ src/libs/database/mysql/connection/MySQLConnection.ts
@@ -3,7 +3,7 @@ import type {
   QueryResult,
   ConnectionConfig,
   PoolStats,
-} from "@/libs/database/interfaces/IDatabaseConnection";
+} from "@/libs/database/interfaces";
 import { ITransaction } from "@/libs/database/interfaces/ITransaction";
 import type { IConnectionAdapter } from "@/libs/database/interfaces/IConnectionAdapter";
 import { createMySQLError } from "../errors";
